fix(jsdoc): final batch - JSDoc warnings and max-lines refactoring
Fixed final 4 JSDoc warnings:
- provider-configs.ts: @default → @defaultValue, escaped braces
- record-format-normalization.test.ts: Added @param/@returns descriptions
- useWebSocket.ts: Removed JSDoc-style type syntax from @returns
- cn.ts: Removed JSDoc-style {type} syntax

Fixed 11 max-lines violations via extraction:
1. base-oauth-provider.test.ts: 865→394 lines (created 3 test files)
2. oauth2-authorization-code.ts: 498→346 lines (created helpers module)
3. browser-adapter.ts: 529→387 lines (created 2 helper modules)
4. node-adapter.integration.test.ts: 417→185 lines (created inspector-helpers)
5. node-adapter.ts: 676→389 lines (created 5 manager modules)
6. scope-inspector.ts: 477→249 lines (created types + utils modules)
7. lightweight-session-manager.ts: 422→365 lines (reused existing factories)
8. session-manager.ts: 747→337 lines (created 3 lifecycle modules)
9. transport-error-factory.ts: 441→356 lines (created HTTP utils)
10. base-client-transport.ts: 492→316 lines (created 4 helper modules)
11. server-connection-manager.ts: 411→379 lines (created manual-ops module)

Result: 0 lint errors, 0 warnings (from 2578 problems)
