+++ packages/auth/src/__tests__/base-oauth-provider.test.ts
@@ -1,16 +1,16 @@
 import { describe, it, expect, beforeEach, vi } from 'vitest';
 
-import { BaseOAuthProvider } from '../implementations/base-oauth-provider.js';
-import {
-  AUTH_DEFAULT_EXPIRY_SECONDS,
-  AUTH_MAX_RETRIES,
-  type OAuth2TokenResponse,
-} from '../utils/index.js';
 import {
   AuthenticationError,
   OAuth2ErrorCode,
 } from '../errors/authentication-error.js';
-import type { ITokenStorage, TokenData } from '@mcp-funnel/core';
+import type { ITokenStorage } from '@mcp-funnel/core';
+import {
+  TestOAuthProvider,
+  MockTokenStorage,
+  createTestToken,
+  createTestTokenResponse,
+} from './base-oauth-provider-test-helpers.js';
 
 // Mock the logger module
 vi.mock('@mcp-funnel/core', async (importOriginal) => {
@@ -20,144 +20,6 @@ vi.mock('@mcp-funnel/core', async (importOriginal) => {
     logEvent: vi.fn(),
   };
 });
-
-// Mock implementation of BaseOAuthProvider for testing
-class TestOAuthProvider extends BaseOAuthProvider {
-  public acquireTokenMock = vi.fn(() => Promise.resolve());
-  public makeTokenRequestMock = vi.fn(() =>
-    Promise.resolve({} as OAuth2TokenResponse),
-  );
-
-  constructor(storage: ITokenStorage) {
-    super(storage);
-  }
-
-  protected async acquireToken(): Promise<void> {
-    return this.acquireTokenMock();
-  }
-
-  // Expose protected methods for testing
-  public async testEnsureValidToken(): Promise<TokenData> {
-    return this.ensureValidToken();
-  }
-
-  public async testRequestTokenWithRetry(
-    makeTokenRequest: () => Promise<OAuth2TokenResponse>,
-    requestId: string,
-  ): Promise<OAuth2TokenResponse> {
-    return this.requestTokenWithRetry(makeTokenRequest, requestId);
-  }
-
-  public async testProcessTokenResponse(
-    tokenResponse: OAuth2TokenResponse,
-    requestId: string,
-    validateAudience?: (audience: string) => boolean,
-  ): Promise<void> {
-    return this.processTokenResponse(
-      tokenResponse,
-      requestId,
-      validateAudience,
-    );
-  }
-
-  public async testHandleTokenRequestError(
-    error: unknown,
-    response?: Response,
-  ): Promise<never> {
-    return this.handleTokenRequestError(error, response);
-  }
-
-  public testValidateTokenResponse(tokenResponse: OAuth2TokenResponse): void {
-    return this.validateTokenResponse(tokenResponse);
-  }
-
-  public testGenerateRequestId(): string {
-    return this.generateRequestId();
-  }
-}
-
-// Mock token storage implementation
-class MockTokenStorage implements ITokenStorage {
-  private token: TokenData | null = null;
-  private refreshCallback?: () => Promise<void> | void;
-
-  public storeMock = vi.fn();
-  public retrieveMock = vi.fn();
-  public isExpiredMock = vi.fn();
-  public clearMock = vi.fn();
-  public scheduleRefreshMock = vi.fn();
-
-  async store(token: TokenData): Promise<void> {
-    this.token = token;
-    return this.storeMock(token);
-  }
-
-  async retrieve(): Promise<TokenData | null> {
-    const result = await this.retrieveMock();
-    return result ?? this.token;
-  }
-
-  async isExpired(): Promise<boolean> {
-    return this.isExpiredMock();
-  }
-
-  async clear(): Promise<void> {
-    this.token = null;
-    return this.clearMock();
-  }
-
-  scheduleRefresh(callback: () => Promise<void> | void): void {
-    this.refreshCallback = callback;
-    this.scheduleRefreshMock(callback);
-  }
-
-  // Helper methods for testing
-  setToken(token: TokenData | null): void {
-    this.token = token;
-  }
-
-  async triggerRefreshCallback(): Promise<void> {
-    if (this.refreshCallback) {
-      await this.refreshCallback();
-    }
-  }
-}
-
-// Helper to create test token data
-/**
- * Creates a test token with specified expiration time.
- *
- * @param expiresInMs - Time in milliseconds until token expires
- * @returns Test token data object
- */
-function createTestToken(expiresInMs: number = 3600000): TokenData {
-  return {
-    accessToken:
-      'test-access-token-' + Math.random().toString(36).substring(2, 11),
-    expiresAt: new Date(Date.now() + expiresInMs),
-    tokenType: 'Bearer',
-    scope: 'read write',
-  };
-}
-
-// Helper to create test OAuth2 token response
-/**
- * Creates a test OAuth2 token response with optional overrides.
- *
- * @param overrides - Partial OAuth2 token response to override defaults
- * @returns Complete OAuth2 token response object
- */
-function createTestTokenResponse(
-  overrides: Partial<OAuth2TokenResponse> = {},
-): OAuth2TokenResponse {
-  return {
-    access_token: 'test-token-' + Math.random().toString(36).substring(2, 11),
-    token_type: 'Bearer',
-    expires_in: 3600,
-    scope: 'read write',
-    ...overrides,
-  };
-}
 
 describe('BaseOAuthProvider', () => {
   let provider: TestOAuthProvider;
@@ -412,202 +274,6 @@ describe('BaseOAuthProvider', () => {
     });
   });
 
-  describe('requestTokenWithRetry', () => {
-    it('should return token response on first successful attempt', async () => {
-      const tokenResponse = createTestTokenResponse();
-      const makeRequest = vi.fn().mockResolvedValue(tokenResponse);
-
-      const result = await provider.testRequestTokenWithRetry(
-        makeRequest,
-        'test-request-id',
-      );
-
-      expect(result).toEqual(tokenResponse);
-      expect(makeRequest).toHaveBeenCalledTimes(1);
-    });
-
-    it('should retry on retryable network errors', async () => {
-      const tokenResponse = createTestTokenResponse();
-      const networkError = new Error('ECONNRESET: Connection reset by peer');
-      const makeRequest = vi
-        .fn()
-        .mockRejectedValueOnce(networkError)
-        .mockRejectedValueOnce(networkError)
-        .mockResolvedValue(tokenResponse);
-
-      const result = await provider.testRequestTokenWithRetry(
-        makeRequest,
-        'test-request-id',
-      );
-
-      expect(result).toEqual(tokenResponse);
-      expect(makeRequest).toHaveBeenCalledTimes(3);
-    });
-
-    it('should not retry on authentication errors', async () => {
-      const authError = new AuthenticationError(
-        'Invalid client',
-        OAuth2ErrorCode.INVALID_CLIENT,
-      );
-      const makeRequest = vi.fn().mockRejectedValue(authError);
-
-      await expect(
-        provider.testRequestTokenWithRetry(makeRequest, 'test-request-id'),
-      ).rejects.toThrow(authError);
-
-      expect(makeRequest).toHaveBeenCalledTimes(1);
-    });
-
-    it('should throw error after maximum retries exceeded', async () => {
-      const networkError = new Error('ETIMEDOUT: Connection timed out');
-      const makeRequest = vi.fn().mockRejectedValue(networkError);
-
-      await expect(
-        provider.testRequestTokenWithRetry(makeRequest, 'test-request-id'),
-      ).rejects.toThrow(networkError);
-
-      expect(makeRequest).toHaveBeenCalledTimes(AUTH_MAX_RETRIES);
-    });
-
-    it('should use exponential backoff for retries', async () => {
-      const networkError = new Error('ECONNREFUSED: Connection refused');
-      const makeRequest = vi.fn().mockRejectedValue(networkError);
-
-      // Mock setTimeout to capture delay values
-      const originalSetTimeout = global.setTimeout;
-      const setTimeoutMock = vi.fn().mockImplementation((fn, _delay) => {
-        fn(); // Execute immediately for testing
-        return {} as NodeJS.Timeout;
-      });
-      global.setTimeout = setTimeoutMock as never;
-
-      try {
-        await expect(
-          provider.testRequestTokenWithRetry(makeRequest, 'test-request-id'),
-        ).rejects.toThrow(networkError);
-
-        // Check exponential backoff delays: 1000ms, 2000ms
-        expect(setTimeoutMock).toHaveBeenCalledTimes(2);
-        expect(setTimeoutMock).toHaveBeenNthCalledWith(
-          1,
-          expect.any(Function),
-          1000,
-        );
-        expect(setTimeoutMock).toHaveBeenNthCalledWith(
-          2,
-          expect.any(Function),
-          2000,
-        );
-      } finally {
-        global.setTimeout = originalSetTimeout;
-      }
-    });
-
-    it('should handle non-Error objects thrown', async () => {
-      const stringError = 'String error';
-      const makeRequest = vi.fn().mockRejectedValue(stringError);
-
-      await expect(
-        provider.testRequestTokenWithRetry(makeRequest, 'test-request-id'),
-      ).rejects.toThrow('String error');
-
-      expect(makeRequest).toHaveBeenCalledTimes(1);
-    });
-  });
-
-  describe('processTokenResponse', () => {
-    it('should store token and schedule refresh', async () => {
-      const tokenResponse = createTestTokenResponse();
-      mockStorage.storeMock.mockResolvedValue(undefined);
-
-      await provider.testProcessTokenResponse(tokenResponse, 'test-request-id');
-
-      expect(mockStorage.storeMock).toHaveBeenCalledWith(
-        expect.objectContaining({
-          accessToken: tokenResponse.access_token,
-          tokenType: tokenResponse.token_type ?? 'Bearer',
-          scope: tokenResponse.scope,
-        }),
-      );
-      expect(mockStorage.scheduleRefreshMock).toHaveBeenCalled();
-    });
-
-    it('should validate audience when provided', async () => {
-      const tokenResponse = createTestTokenResponse({
-        audience: 'https://api.example.com',
-      });
-      const validateAudience = vi.fn().mockReturnValue(true);
-
-      await provider.testProcessTokenResponse(
-        tokenResponse,
-        'test-request-id',
-        validateAudience,
-      );
-
-      expect(validateAudience).toHaveBeenCalledWith('https://api.example.com');
-      expect(mockStorage.storeMock).toHaveBeenCalled();
-    });
-
-    it('should throw error for invalid audience', async () => {
-      const tokenResponse = createTestTokenResponse({
-        audience: 'https://wrong.example.com',
-      });
-      const validateAudience = vi.fn().mockReturnValue(false);
-
-      await expect(
-        provider.testProcessTokenResponse(
-          tokenResponse,
-          'test-request-id',
-          validateAudience,
-        ),
-      ).rejects.toThrow(AuthenticationError);
-      await expect(
-        provider.testProcessTokenResponse(
-          tokenResponse,
-          'test-request-id',
-          validateAudience,
-        ),
-      ).rejects.toThrow('Audience validation failed');
-    });
-
-    it('should continue even if storage fails', async () => {
-      const tokenResponse = createTestTokenResponse();
-      mockStorage.storeMock.mockRejectedValue(new Error('Storage failed'));
-
-      await expect(
-        provider.testProcessTokenResponse(tokenResponse, 'test-request-id'),
-      ).resolves.not.toThrow();
-
-      expect(mockStorage.storeMock).toHaveBeenCalled();
-    });
-
-    it('should use default expiry when not provided', async () => {
-      const tokenResponse = createTestTokenResponse({ expires_in: undefined });
-
-      await provider.testProcessTokenResponse(tokenResponse, 'test-request-id');
-
-      const storedToken = mockStorage.storeMock.mock.calls[0]?.[0];
-      expect(storedToken.expiresAt.getTime()).toBeCloseTo(
-        Date.now() + AUTH_DEFAULT_EXPIRY_SECONDS * 1000,
-        -3,
-      );
-    });
-
-    it('should skip audience validation when not provided', async () => {
-      const tokenResponse = createTestTokenResponse({ audience: undefined });
-      const validateAudience = vi.fn();
-
-      await provider.testProcessTokenResponse(
-        tokenResponse,
-        'test-request-id',
-        validateAudience,
-      );
-
-      expect(validateAudience).not.toHaveBeenCalled();
-      expect(mockStorage.storeMock).toHaveBeenCalled();
-    });
-  });
-
   describe('handleTokenRequestError', () => {
     it('should re-throw AuthenticationError directly', async () => {
       const authError = new AuthenticationError(
@@ -721,145 +387,6 @@ describe('BaseOAuthProvider', () => {
 
       const uuidRegex = /^\d{13}_[a-f0-9]{8}$/i;
       expect(id).toMatch(uuidRegex);
-    });
-  });
-
-  describe('Proactive Refresh Scheduling', () => {
-    it('should schedule refresh for tokens with sufficient time', async () => {
-      const token = createTestToken(3600000); // 1 hour
-      mockStorage.setToken(token);
-      mockStorage.isExpiredMock.mockResolvedValue(false);
-
-      await provider.testEnsureValidToken();
-
-      expect(mockStorage.scheduleRefreshMock).toHaveBeenCalled();
-      const callback = mockStorage.scheduleRefreshMock.mock.calls[0][0];
-      expect(typeof callback).toBe('function');
-    });
-
-    it('should not schedule refresh for tokens expiring soon', async () => {
-      const token = createTestToken(60000); // 1 minute (less than 5 minute buffer)
-      mockStorage.setToken(token);
-      mockStorage.isExpiredMock.mockResolvedValue(false);
-
-      await provider.testEnsureValidToken();
-
-      // ScheduleRefresh should NOT be called because refresh time would be in the past
-      expect(mockStorage.scheduleRefreshMock).not.toHaveBeenCalled();
-    });
-
-    it('should handle refresh callback errors gracefully', async () => {
-      const token = createTestToken(3600000);
-      mockStorage.setToken(token);
-      mockStorage.isExpiredMock.mockResolvedValue(false);
-
-      // Mock refresh to fail
-      provider.acquireTokenMock.mockRejectedValue(new Error('Refresh failed'));
-
-      await provider.testEnsureValidToken();
-
-      // Simulate proactive refresh callback
-      await expect(mockStorage.triggerRefreshCallback()).resolves.not.toThrow();
-    });
-  });
-
-  describe('Integration Scenarios', () => {
-    it('should handle full token lifecycle', async () => {
-      // Initial state - no token
-      mockStorage.setToken(null);
-
-      // First call should acquire token
-      const newToken = createTestToken();
-      provider.acquireTokenMock.mockImplementation(async () => {
-        mockStorage.setToken(newToken);
-      });
-      mockStorage.isExpiredMock.mockResolvedValue(false);
-
-      const headers1 = await provider.getHeaders();
-      expect(headers1.Authorization).toBe(`Bearer ${newToken.accessToken}`);
-      expect(provider.acquireTokenMock).toHaveBeenCalledTimes(1);
-
-      // Second call should use cached token
-      const headers2 = await provider.getHeaders();
-      expect(headers2.Authorization).toBe(`Bearer ${newToken.accessToken}`);
-      expect(provider.acquireTokenMock).toHaveBeenCalledTimes(1);
-
-      // Token expires
-      mockStorage.isExpiredMock.mockResolvedValue(true);
-      const refreshedToken = createTestToken();
-      provider.acquireTokenMock.mockImplementation(async () => {
-        mockStorage.setToken(refreshedToken);
-      });
-
-      // Third call should refresh token
-      const headers3 = await provider.getHeaders();
-      expect(headers3.Authorization).toBe(
-        `Bearer ${refreshedToken.accessToken}`,
-      );
-      expect(provider.acquireTokenMock).toHaveBeenCalledTimes(2);
-    });
-
-    it('should handle concurrent requests during token acquisition', async () => {
-      mockStorage.setToken(null);
-
-      let resolveAcquireToken: () => void;
-      const acquireTokenPromise = new Promise<void>((resolve) => {
-        resolveAcquireToken = resolve;
-      });
-
-      const newToken = createTestToken();
-      provider.acquireTokenMock.mockImplementation(async () => {
-        await acquireTokenPromise;
-        mockStorage.setToken(newToken);
-      });
-
-      // Start multiple concurrent requests
-      const headerPromises = [
-        provider.getHeaders(),
-        provider.getHeaders(),
-        provider.getHeaders(),
-      ];
-
-      // Give a small delay to allow promises to start
-      await new Promise((resolve) => setTimeout(resolve, 10));
-
-      // Token acquisition should be in progress
-      expect(provider.acquireTokenMock).toHaveBeenCalledTimes(1);
-
-      // Complete token acquisition
-      resolveAcquireToken!();
-
-      const results = await Promise.all(headerPromises);
-
-      // All requests should get the same token
-      results.forEach((headers) => {
-        expect(headers.Authorization).toBe(`Bearer ${newToken.accessToken}`);
-      });
-
-      // Should only acquire token once
-      expect(provider.acquireTokenMock).toHaveBeenCalledTimes(1);
-    });
-
-    it('should handle storage errors gracefully', async () => {
-      const token = createTestToken();
-      mockStorage.setToken(token);
-
-      // First call to retrieve fails, then succeeds after refresh
-      mockStorage.retrieveMock
-        .mockRejectedValueOnce(new Error('Storage error'))
-        .mockResolvedValue(token);
-
-      // Should still be able to operate
-      const isValid = await provider.isValid();
-      expect(isValid).toBe(false);
-
-      // Should try to acquire new token
-      provider.acquireTokenMock.mockImplementation(async () => {
-        mockStorage.setToken(token);
-      });
-
-      const headers = await provider.getHeaders();
-      expect(headers.Authorization).toBe(`Bearer ${token.accessToken}`);
     });
   });
 });
